requirements.txt
# ============================================
# Voice-to-Voice AI Assistant for Product Discovery
# E-commerce Final Project
# ============================================

# ===================
# MCP Server
# ===================
fastmcp>=0.4.0
httpx>=0.25.0
pydantic>=2.0.0
aiohttp>=3.9.0

# ===================
# Vector Store & RAG
# ===================
chromadb>=0.4.0
sentence-transformers>=2.2.0

# ===================
# LLM Clients (model-agnostic)
# ===================
openai>=1.0.0
# anthropic>=0.18.0        # Uncomment for Anthropic/Claude
# google-generativeai>=0.3.0  # Uncomment for Google Gemini

# ===================
# Data Processing
# ===================
pandas>=2.0.0
numpy<2.0.0  # Important: numpy 2.x has compatibility issues with torch
pyarrow>=14.0.0

# ===================
# ML/Deep Learning
# ===================
torch>=2.0.0

# ===================
# Environment & Utilities
# ===================
python-dotenv>=1.0.0

# ===================
# Testing (optional)
# ===================
pytest>=7.0.0
pytest-asyncio>=0.21.0

opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-proto==1.37.0
